LayoutTests/fast/frames/frame-focus-blurs-active-element-expected.txt

CONSOLE MESSAGE: PASS: blur event was fired on the active element when the iframe was focused.
